int -> unsigned int
`size_t' is used to represent the size of an object (i.e. size_t is unsigned). It is platform-dependent.
`int'  is signed by default, and it is fixed 32bits on most modern machine.
So in 32-bit machine: size_t is 32-bit, unsigned int is 32-bit.
But in 64-bit machine: size_t is 64-bit, unsigned int is still 32-bit.